저글러 3-11특성: 적 이외의 대상에게 투척 무기를 발사해도 골드가 떨어지는 문제 수정

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Juggling.java
@@ -169,7 +169,7 @@ public class Juggling extends Buff implements ActionIndicator.Action {
                                 public void call() {
                                     if (Dungeon.hero.hasTalent(Talent.FANCY_PERFORMANCE)) {
                                         Char ch = Actor.findChar(destination);
-                                        if (ch != null) {
+                                        if (ch != null && ch.alignment == Char.Alignment.ENEMY) {
                                             Dungeon.level.drop(new Gold(5*Dungeon.hero.pointsInTalent(Talent.FANCY_PERFORMANCE)), destination).sprite.drop();
                                         }
                                     }
